final clean up of sha code
